requirements.txt
minio==7.2.15
dotenv==0.9.9
boto3==1.37.23
pillow==11.1.0
django-filter==25.1
drf-spectacular==0.28.0
django-storages==1.14.5
psycopg2-binary==2.9.10
djangorestframework==3.16.0
drf-spectacular-sidecar==2025.3.1